+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Data/GitHubLicense.cs
@@ -1,16 +1,17 @@
-using System.Collections;
-using System.Collections.Generic;
-using Codice.CM.SemanticMerge.Gui;
-using UnityEngine;
+using Newtonsoft.Json;
 
 namespace Studio23.SS2.UPMAssistant.Editor.Data
 {
-    [System.Serializable]
+    [JsonObject(MemberSerialization.OptIn)]
     public class GitHubLicense
     {
-        public string key;
-        public string name;
-        public string url;
+        [JsonProperty("key")]
+        public string Key { get; set; }
+        
+        [JsonProperty("name")]
+        public string Name { get; set; }
+        
+        [JsonProperty("url")]
+        public string URL { get; set; }
     }
 }
-