+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/FileEditorWindowController.cs
@@ -6,17 +6,17 @@ namespace Studio23.SS2.UPMAssistant.Editor
 {
     public class FileEditorWindowController : EditorWindow
     {
-        private static string markupText = ""; 
-        private static string filePath = "";  
+        private static string _markupText = ""; 
+        private static string _filePath = "";  
         
         public static void ShowWindow(string path)
         {
             if (!string.IsNullOrEmpty(path))
             {
-                filePath = path;
+                _filePath = path;
                 var window =GetWindow<FileEditorWindowController>("UPM Editor Window");
                 window.minSize = new Vector2(600, 600); 
-                markupText = File.ReadAllText(filePath);
+                _markupText = File.ReadAllText(_filePath);
             }
             else
             {
@@ -27,18 +27,18 @@ namespace Studio23.SS2.UPMAssistant.Editor
         private void OnGUI()
         {
             GUILayout.Label("UPM Editor", EditorStyles.boldLabel);
-            GUILayout.Label($"File Location: {filePath}", EditorStyles.label);
+            GUILayout.Label($"File Location: {_filePath}", EditorStyles.label);
             
-            markupText = EditorGUILayout.TextArea(markupText, GUILayout.Height(500));
+            _markupText = EditorGUILayout.TextField(_markupText, GUILayout.Height(500));
              
-            var btnStatus = !string.IsNullOrEmpty(filePath) && !string.IsNullOrEmpty(markupText);
+            var btnStatus = !string.IsNullOrEmpty(_filePath) && !string.IsNullOrEmpty(_markupText);
             EditorGUI.BeginDisabledGroup(!btnStatus);
             GUI.backgroundColor = Color.green;
             if (GUILayout.Button("Save", GUILayout.Height(40)))
             {
-                if (!string.IsNullOrEmpty(filePath))
+                if (!string.IsNullOrEmpty(_filePath))
                 {
-                    File.WriteAllText(filePath, markupText);
+                    File.WriteAllText(_filePath, _markupText);
                     AssetDatabase.Refresh();
                 }
                 else
@@ -48,14 +48,6 @@ namespace Studio23.SS2.UPMAssistant.Editor
             }
             GUI.backgroundColor = Color.white; // Reset the background color
             EditorGUI.EndDisabledGroup();
-            
-             
-             
-           
-            
         }
-
-         
     }
-
 }