+++ Assets/Packages/com.studio23.ss2.upmassistant/Editor/Utility/SharedGUIContent.cs
@@ -11,8 +11,6 @@ namespace Studio23.SS2.UPMAssistant.Editor.Data
     public class SharedGUIContent : ScriptableObject
     {
         private static SharedGUIContent _instance;
-
-        
         public static SharedGUIContent Instance
         {
             get
@@ -32,10 +30,10 @@ namespace Studio23.SS2.UPMAssistant.Editor.Data
           
             #region Licenses
 
-            if (APIHandler.gitHubLicense != null && APIHandler.gitHubLicense.Count > 0)
+            if (APIHandler.GitHubLicense != null && APIHandler.GitHubLicense.Count > 0)
             {
                 List<string> licenseNames = new List<string>();
-                foreach (var license in APIHandler.gitHubLicense)  licenseNames.Add(license.name);
+                foreach (var license in APIHandler.GitHubLicense)  licenseNames.Add(license.Name);
 
                 APIHandler.SelectedLicenceIndex = EditorGUILayout.Popup("Select License", APIHandler.SelectedLicenceIndex, licenseNames.ToArray());
                 
@@ -44,7 +42,7 @@ namespace Studio23.SS2.UPMAssistant.Editor.Data
                 if (GUILayout.Button("Configure"))
                 {
                     var licenseFilePath =
-                        Path.Combine(DataHandler.ROOT + DataHandler.LoadPackageNameData(), DataHandler.LICENSE_MD);
+                        Path.Combine(DataHandler.Root + DataHandler.GetSavedPackagedName(), DataHandler.LicenseMD);
                     if (File.Exists(licenseFilePath))
                         FileEditorWindowController.ShowWindow(licenseFilePath);
                     else  Debug.LogError("License file not found!");